+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 CHANGELOG
 =========
+0.0.5 (2023-07-21)
 
+-added timeout option with default 5000ms to prevent DOS attacks and bad clients not 
+ closing sockets properly. this is mainly used for server connected directly to NLB's
 
 0.0.1 (2023-07-17)